@@ -4,7 +4,9 @@ public class PixivUrl {
     static final String BASE_URL = "https://www.pixiv.net";
     static final String ARTWORK_URL = "https://www.pixiv.net/artworks/";
     static final String WEEKLY_URL = "https://www.pixiv.net/ranking.php?mode=weekly&content=illust";
-    static final String MONTHLY_URL = "https://www.pixiv.net/ranking.php?mode=monthly&content=illust";
+    static final String MONTHLY_RANKING_ILLUST_URL = "https://www.pixiv.net/ranking.php?mode=monthly&content=illust";
+    static final String MALE_RANKING = "https://www.pixiv.net/ranking.php?mode=male";
+    static final String USER_PAGE = "https://www.pixiv.net/users/";
     static final String LOGIN_PAGE_URL = "https://accounts.pixiv.net/login?lang=zh&source=pc&view_type=page&ref=wwwtop_accounts_index";
     static final String LOGIN_API_URL = "https://accounts.pixiv.net/api/login?lang=zh";
     static final String[] User_Agents = new String[] {
